Move modules common to ui and server to a separate folder
Move game specific code and tests under separate folders.
