+++ lib/common/engine.js
