+++ lib/common/jkiss.js
