+++ lib/common/panel-league/basic.js
