+++ lib/common/panel-league/garbage.js
