+++ lib/common/panel-league/scoring.js
