+++ lib/common/panel-league/stepper.js
@@ -4,7 +4,7 @@ const basic = require('./basic');
 const garbage = require('./garbage');
 const defaultTo = require('lodash/defaultTo');
 
-const getScore = require('../scoring');
+const getScore = require('./scoring');
 
 
 const defaultOptions = {
