+++ lib/common/panel-league/util.js
