+++ lib/server/game.js
@@ -2,7 +2,7 @@
 
 const uuidV4 = require('uuid/v4');
 
-const { GameEngine } = require('../engine');
+const { GameEngine } = require('../common/engine');
 
 
 class Game {
