+++ lib/ui/panel-league/game-duel.js
@@ -5,7 +5,7 @@ import { el, setChildren } from 'redom';
 import Grid from './grid';
 import PanelLeagueGameBase from './game-base';
 
-import { NetworkGameEngine } from '../../engine';
+import { NetworkGameEngine } from '../../common/engine';
 
 
 export default class PanelLeagueGameDuel extends PanelLeagueGameBase {
