+++ test/panel-league/test-garbage.js
@@ -1,6 +1,6 @@
 const times = require('lodash/times');
-const {GameEngine} = require('../lib/engine');
-const {R, G, B, _} = require('../lib/panel-league/util');
+const {GameEngine} = require('../../lib/common/engine');
+const {R, G, B, _} = require('../../lib/common/panel-league/util');
 
 module.exports.testAdd = ((test) => {
   const game = new GameEngine();
